Improve input and error handling in npm_globals.sh
- Only prompt for user input to create symlink if doesn't already exist
- Echo error message if npm package not found, instead of exiting script

+++ bin/npm_globals.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-### Install global npm packages ###
+### --------------------- Install global npm packages --------------------- ###
 # Based on https://github.com/ianwalter/dotnpm/blob/master/install.sh
 # Designed for use with Strap:
 # https://github.com/MikeMcQuaid/strap
@@ -7,11 +7,13 @@
 npm_install_globals() {
   if [ -d ~/.dotfiles ]; then
     (
-      echo "Dotfiles directory found. Installing global npm packages..."
-      # Make directories for symlinks, if they don't already exist
+      echo "-> Dotfiles directory found. Installing global npm packages..."
+      # Make directory for symlink, if it doesh't already exist (-p)
       mkdir -p ~/.npm
-      # Create symlink to package list (ln -s), and prompt (-i) before overwrite (-f)
-      ln -s -f -i ~/.dotfiles/.npm/npm-globals.txt ~/.npm/npm-globals.txt
+      # Create symlink (ln -s), and prompt (-i) before overwrite (-f)
+      if [ ! -f ~/.npm/npm-globals.txt ]; then
+        ln -s -f -i ~/.dotfiles/.npm/npm-globals.txt ~/.npm/npm-globals.txt
+      fi
       # Install npm packages listed in npm-globals.txt
       package_dir="$(npm config get prefix)/lib"
       packages=$(npm ls -g --parseable --depth=0)
@@ -21,19 +23,19 @@ npm_install_globals() {
         installed=$(echo "$packages" | grep -ce "^$p\$")
         if [ "$installed" == "0" ]; then
           echo "Installing $p."
-          npm install -g "$p"
+          npm install -g "$p" || echo "-> Error: package $p not found in npm."
         else
           echo "$p already installed."
         fi
       done <~/.dotfiles/.npm/npm-globals.txt
-      echo "Done installing npm packages."
+      echo "-> Done installing npm packages."
     )
   else
-    echo "Dotfiles directory not found. Installation of npm packages not successful."
+    echo "-> Error: dotfiles directory not found."
   fi
 }
 if npm_install_globals; then
-  echo "npm_install_globals() ran successfully."
+  echo "-> npm_install_globals() ran successfully."
 else
-  echo "npm_install_globals() did not run successfully."
+  echo "-> Error: npm_install_globals() did not run successfully."
 fi
